MILESTONE: Successfully set up inter-process communication
I use preload.js to expose ipcRenderer.send and ipcRenderer.sendSync in my renderer.js. When I invoke these methods, they call main.js and are handled. There is a configModule.js registered in main.js that has an Invoke method that can read the message sent to main.js and pass it on to configModule.js to process. Methods invoked can be synchronous or asynchronous and will work fine as long as the right contextBridge method is invoked to call them.