Updated options save
Added toastTimeout setting to settings page and improved the way settings are loaded and saved (values were each being loaded with a separate call to the configModule.js and saved with separate calls to the configModule.js). Improved layout of the options pages. Fixed an error on reading all user prefs.