Add statusbar and split worldModule into two scripts
Added a statusbar <div> to every page that can display a status message with the option to persist a status from one page to the next

Split worldModule.js into worldModule.js and worldDbModule.js. There will be enough db calls that I need to make that it will be worth having these in two different scripts.