Finished saving outgoing links to database on page save and fixed HTML to Markdown conversion
When saving a page, outgoing links were being added to the database, but never removed, leaving outdated links in the database. I added a line to remove existing links from the database before saving new links.

When saving an MD page, the outgoing links were not being saved to the database. I added the HTML version of the page to the function to allow it to find and log links.

When converting an HTML page to Markdown, the <link rel="stylesheet"/> references were included. I remove <link> tags from the HTML before injecting it into the editor now.